Move User model, repository, service & config in a dedicated package
