+++ backend/src/main/java/com/cartowiki/webapp/authentication/controller/AuthenticationController.java
@@ -13,10 +13,10 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.cartowiki.webapp.authentication.model.User;
-import com.cartowiki.webapp.authentication.model.request.LogInRequest;
-import com.cartowiki.webapp.authentication.model.request.SignUpRequest;
-import com.cartowiki.webapp.authentication.service.UserService;
+import com.cartowiki.webapp.users.model.User;
+import com.cartowiki.webapp.users.service.UserService;
+import com.cartowiki.webapp.authentication.model.LogInRequest;
+import com.cartowiki.webapp.authentication.model.SignUpRequest;
 import com.cartowiki.webapp.authentication.service.JwtService;
 import com.cartowiki.webapp.util.ResponseMaker;
