+++ backend/src/main/java/com/cartowiki/webapp/authentication/model/SignUpRequest.java
@@ -1,4 +1,4 @@
-package com.cartowiki.webapp.authentication.model.request;
+package com.cartowiki.webapp.authentication.model;
 
 /**
  * Data template for sign up requests
