+++ backend/src/main/java/com/cartowiki/webapp/users/config/DatabaseConfig.java
@@ -1,4 +1,4 @@
-package com.cartowiki.webapp.authentication.config;
+package com.cartowiki.webapp.users.config;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
