+++ backend/src/main/java/com/cartowiki/webapp/users/repository/UserRepository.java
@@ -1,11 +1,11 @@
-package com.cartowiki.webapp.authentication.repository;
+package com.cartowiki.webapp.users.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import com.cartowiki.webapp.authentication.model.User;
+import com.cartowiki.webapp.users.model.User;
 
 /**
  * Requests management about the User entity in the database
