+++ backend/src/test/java/com/cartowiki/webapp/authentication/controller/AuthenticationControllerTests.java
@@ -14,9 +14,9 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 
-import com.cartowiki.webapp.authentication.config.DatabaseConfig;
-import com.cartowiki.webapp.authentication.model.User;
-import com.cartowiki.webapp.authentication.repository.UserRepository;
+import com.cartowiki.webapp.users.config.DatabaseConfig;
+import com.cartowiki.webapp.users.model.User;
+import com.cartowiki.webapp.users.repository.UserRepository;
 
 /**
  * Tests requests for authentication
