+++ backend/src/test/java/com/cartowiki/webapp/authentication/filter/JwtAuthFilterTests.java
@@ -12,9 +12,9 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 
-import com.cartowiki.webapp.authentication.model.User;
-import com.cartowiki.webapp.authentication.repository.UserRepository;
 import com.cartowiki.webapp.authentication.service.JwtService;
+import com.cartowiki.webapp.users.model.User;
+import com.cartowiki.webapp.users.repository.UserRepository;
 
 /**
  * Test the filter chain element checking JWT authentication
