+++ backend/src/test/java/com/cartowiki/webapp/users/model/UserTests.java
@@ -1,4 +1,4 @@
-package com.cartowiki.webapp.authentication.model;
+package com.cartowiki.webapp.users.model;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -6,6 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.security.core.GrantedAuthority;
+
 
 /**
  * Tests for User
